+++ src/components/dialog/index.ts
@@ -1,3 +1,5 @@
 export { default } from "./dialog.component";
 export type { DialogProps, DialogHandle } from "./dialog.component";
 export type { DialogSizes } from "./dialog.config";
+export { DialogHeader, DialogHeadingStatus } from "./dialog.component";
+export type { DialogHeadingStatusProps } from "./dialog.component";